requirements.txt
openai
dotenv
python-chess